Language.Preprocessor.Unlit

Description

Part of this code is from Report on the Programming Language Haskell, version 1.2, appendix C.

Documentation

unlit :: FilePath -> String -> String

unlit takes a filename (for error reports), and transforms the given string, to eliminate the literate comments from the program text.